requirements.txt
streamlit>=1.39
pandas>=2.2
numpy>=2.1
pyarrow>=18.0
duckdb>=1.1
plotly>=5.24
pillow>=10.4
pytest>=8.3

